requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astronomer-cosmos == 1.6
apache-airflow-providers-postgres==5.6.0

# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-airbyte==3.8.1
apache-airflow-providers-docker==3.13.0
apache-airflow-providers-http==4.12.0
astronomer-providers==1.19.2
dbt-adapters==1.2.1
dbt-common==1.2.0
dbt-core==1.8.1
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
dbt-postgres==1.8.2
gspread==6.1.2
psycopg2-binary==2.9.9
oauth2client==4.1.3